requirements.txt
environs==5.2.1
python-telegram-bot==13.7
requests==2.26.0
tldextract==3.1.2